+++ app/src/main/java/com/netmera/netmerafintech/Netmera/Events/ForgotYourPinEvent.kt
@@ -1,3 +1,7 @@
+/*
+* Copyright (c) 2023 Netmera.
+*/
+
 package com.netmera.netmerafintech.Netmera.Events
 
 import com.netmera.NetmeraEvent
